Refactor category_dao_impl naming
